Re-autogenerate bindings
This additionally commits a script to re-generate the bindings
